[kexu] Adding caching of dependencies to github action to speed up
In the future, we might need to switch caching away from the `node_modules` folders back to the `~/.npm' folder.
The `node_modules` caching is a trick we use, if we assume our `package-lock.json` file won't change super frequently.
If there's a change in the package lock, we won't revert back to a default cache, we'll make a new cache.